Modified Rectangle assembly program for better demo
Originally, the the rectangle program makes a rectangle the length that you put into register R0. At first, I didn't realize this, and I was confused as to why there was no rectangle appearing because R0 was 0. Therefore, for better demonstration purposes, I removed this input feature because I was likely to forget about it during the demo. In order to do so, I loaded a constant into the address register, and then set the D register to the A register.